fix types in KeyConditionIndexStat commit_hash:a1a96330e83a4d5708b22c89a510b4a8e8424630

+++ yt/chyt/server/index_stats.h
@@ -24,7 +24,7 @@ struct IChytIndexStat
 
 std::shared_ptr<IChytIndexStat> CreateVirtualColumnIndexStat(int discardedTableCount, int inputTablesCount);
 
-std::shared_ptr<IChytIndexStat> CreateKeyConditionIndexStat(int rowCount, int dataWeight, int filteredRowCount, int filteredDataWeight);
+std::shared_ptr<IChytIndexStat> CreateKeyConditionIndexStat(i64 rowCount, i64 dataWeight, i64 filteredRowCount, i64 filteredDataWeight);
 
 ////////////////////////////////////////////////////////////////////////////////
